+++ tests/data/unit_tests/test_merger/test_merge_lines_split_by_operators.sql
@@ -1,0 +1,65 @@
+select
+    a_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field,
+    b_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field,
+    c_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field,
+    d_field,
+    e_field
+where
+    something
+    < something_else
+    and another_thing
+    = another_thing_entirely
+    or maybe_something_completely_different
+)))))__SQLFMT_OUTPUT__(((((
+select
+    a_field + another_field + another_field + another_field + another_field,
+    b_field + another_field + another_field + another_field + another_field,
+    c_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field
+    + another_field,
+    d_field,
+    e_field
+where
+    something < something_else
+    and another_thing = another_thing_entirely
+    or maybe_something_completely_different